docs: complete DRY improvements for test mocking patterns in Story 2.2
## CHANGES

- Extract common test mocking patterns to shared utilities
- Create FabricApiMockBuilder with fluent API
- Consolidate port management functions
- Refactor 8 test files to use shared utilities
- Reduce code duplication from 6+ to 2 instances
- Improve pylint score to 9.99/10
- Maintain all 143 tests passing
