+++ tests/shared/port_utils.py
@@ -1,0 +1,22 @@
+"""Shared port utilities for testing."""
+
+import socket
+
+
+def find_free_port() -> int:
+    """Find a free port for testing purposes."""
+    with socket.socket(socket.AF_INET, socket.SOCK_STREAM) as s:
+        s.bind(("", 0))
+        s.listen(1)
+        port = s.getsockname()[1]
+    return port
+
+
+def is_port_in_use(port: int, host: str = "127.0.0.1") -> bool:
+    """Check if a port is already in use."""
+    with socket.socket(socket.AF_INET, socket.SOCK_STREAM) as s:
+        try:
+            s.connect((host, port))
+            return True
+        except (OSError, ConnectionRefusedError):
+            return False
